app/requirements.txt
requests
Flask
python-dotenv